requirements.txt
fastapi
uvicorn
seedrcc
python-dotenv
requests
